+++ examples/Cargo.toml
@@ -6,7 +6,7 @@ publish = false
 
 [dependencies]
 vulkano = { path = "../vulkano" }
-vulkano-shader-derive = { path = "../vulkano-shader-derive" }
+vulkano-shaders = { path = "../vulkano-shaders" }
 vulkano-win = { path = "../vulkano-win" }
 cgmath = "0.16.1"
 image = "0.20.0"
