+++ examples/src/bin/deferred/frame/point_lighting_system.rs
@@ -23,7 +23,7 @@ use vulkano::pipeline::blend::BlendOp;
 use vulkano::pipeline::GraphicsPipeline;
 use vulkano::pipeline::GraphicsPipelineAbstract;
 use vulkano::pipeline::viewport::Viewport;
-use vulkano_shader_derive::vulkano_shader;
+use vulkano_shaders::vulkano_shader;
 use cgmath::Matrix4;
 use cgmath::Vector3;
