+++ examples/src/bin/deferred/main.rs
@@ -31,7 +31,7 @@
 extern crate cgmath;
 #[macro_use]
 extern crate vulkano;
-extern crate vulkano_shader_derive;
+extern crate vulkano_shaders;
 extern crate winit;
 extern crate vulkano_win;
