+++ examples/src/bin/tessellation.rs
@@ -22,7 +22,7 @@
 
 #[macro_use]
 extern crate vulkano;
-extern crate vulkano_shader_derive;
+extern crate vulkano_shaders;
 extern crate winit;
 extern crate vulkano_win;
 
@@ -46,7 +46,7 @@ use vulkano::swapchain::AcquireError;
 use vulkano::swapchain::SwapchainCreationError;
 use vulkano::sync::now;
 use vulkano::sync::GpuFuture;
-use vulkano_shader_derive::vulkano_shader;
+use vulkano_shaders::vulkano_shader;
 
 use std::sync::Arc;
