+++ vulkano-shader-derive/Cargo.toml
@@ -1,22 +1,9 @@
 [package]
 name = "vulkano-shader-derive"
-version = "0.10.0"
+version = "0.11.0"
 authors = ["Pierre Krieger <pierre.krieger1708@gmail.com>", "The vulkano contributors"]
 repository = "https://github.com/vulkano-rs/vulkano"
-description = "Safe wrapper for the Vulkan graphics API"
+description = "Deprecated"
 license = "MIT/Apache-2.0"
 documentation = "https://docs.rs/vulkano"
-categories = ["rendering::graphics-api"]
-
-[lib]
-name = "vulkano_shader_derive"
-proc-macro = true
-
-[dependencies]
-syn = "0.15"
-quote = "0.6"
-proc-macro2 = "0.4"
-vulkano-shaders = { version = "0.10", path = "../vulkano-shaders" }
-
-[dev-dependencies]
-vulkano = { version = "0.10", path = "../vulkano" }
+categories = []
