+++ vulkano-shaders/src/lib.rs
@@ -1,32 +1,179 @@
-// Copyright (c) 2016 The vulkano developers
-// Licensed under the Apache License, Version 2.0
-// <LICENSE-APACHE or
-// http://www.apache.org/licenses/LICENSE-2.0> or the MIT
-// license <LICENSE-MIT or http://opensource.org/licenses/MIT>,
-// at your option. All files in the project carrying such
-// notice may not be copied, modified, or distributed except
-// according to those terms.
+//! Procedural Macro glue for compile-time compilation of GLSL into SPIR-V
+//!
+//! # Basic usage
+//!
+//! ```
+//! #[macro_use]
+//! extern crate vulkano_shader_derive;
+//! extern crate vulkano;
+//! # fn main() {}
+//! #[allow(unused)]
+//! mod vertex_shader {
+//!     #[derive(VulkanoShader)]
+//!     #[ty = "vertex"]
+//!     #[src = "
+//! #version 450
+//!
+//! layout(location = 0) in vec3 position;
+//!
+//! void main() {
+//!     gl_Position = vec4(position, 1.0);
+//! }
+//! "]
+//!     struct Dummy;
+//! }
+//! ```
+//!
+//! # Details
+//!
+//! Due to the current limitations of procedural shaders in Rust, the current
+//! functionality of this crate is to base everything off of deriving
+//! `VulkanoShader` for a dummy struct that never actually gets used. When
+//! derived, the unused struct itself will be replaced by the functionality
+//! needed to use the shader in a Vulkano application. Due to the fact that
+//! a lot of what is generated will never be used, it's a good idea to put
+//! `#[allow(unused)]` on the module itself if you don't want to see irrelevant
+//! errors.
+//! 
+//! If you want to take a look at what the macro generates, your best options
+//! are to either read through the code that handles the generation (the
+//! [`reflect`][reflect] function in the `vulkano-shaders` crate) or use a tool
+//! such as [cargo-expand][cargo-expand] to view the expansion of the macro in your
+//! own code. It is unfortunately not possible to provide a `generated_example`
+//! module like some normal macro crates do since derive macros cannot be used from
+//! the crate they are declared in. On the other hand, if you are looking for a
+//! high-level overview, you can see the below section.
+//! 
+//! # Generated code overview
+//! 
+//! The macro generates the following items of interest:
+//! * The `Shader` struct. This contains a single field, `shader`, which is an
+//! `Arc<ShaderModule>`.
+//! * The `Shader::load` constructor. This method takes an `Arc<Device>`, calls
+//! [`ShaderModule::new`][ShaderModule::new] with the passed-in device and the
+//! shader data provided via the macro, and returns `Result<Shader, OomError>`.
+//! Before doing so, it loops through every capability instruction in the shader
+//! data, verifying that the passed-in `Device` has the appropriate features
+//! enabled. **This function currently panics if a feature required by the shader
+//! is not enabled on the device.** At some point in the future it will return
+//! an error instead.
+//! * The `Shader::module` method. This method simply returns a reference to the
+//! `Arc<ShaderModule>` contained within the `shader` field of the `Shader`
+//! struct.
+//! * Methods for each entry point of the shader module. These construct and
+//! return the various entry point structs that can be found in the
+//! [vulkano::pipeline::shader][pipeline::shader] module.
+//! * A Rust struct translated from each struct contained in the shader data.
+//! * The `Layout` newtype. This contains a [`ShaderStages`][ShaderStages] struct.
+//! An implementation of [`PipelineLayoutDesc`][PipelineLayoutDesc] is also
+//! generated for the newtype.
+//! * The `SpecializationConstants` struct. This contains a field for every
+//! specialization constant found in the shader data. Implementations of
+//! `Default` and [`SpecializationConstants`][SpecializationConstants] are also
+//! generated for the struct.
+//! 
+//! All of these generated items will be accessed through the module that you
+//! wrote to use the derive macro in. If you wanted to store the `Shader` in
+//! a struct of your own, you could do something like this:
+//! 
+//! ```
+//! # #[macro_use]
+//! # extern crate vulkano_shader_derive;
+//! # extern crate vulkano;
+//! # fn main() {}
+//! # use std::sync::Arc;
+//! # use vulkano::OomError;
+//! # use vulkano::device::Device;
+//! #
+//! # #[allow(unused)]
+//! # mod vertex_shader {
+//! #     #[derive(VulkanoShader)]
+//! #     #[ty = "vertex"]
+//! #     #[src = "
+//! # #version 450
+//! #
+//! # layout(location = 0) in vec3 position;
+//! #
+//! # void main() {
+//! #     gl_Position = vec4(position, 1.0);
+//! # }
+//! # "]
+//! #     struct Dummy;
+//! # }
+//! // various use statements
+//! // `vertex_shader` module with shader derive
+//! 
+//! pub struct Shaders {
+//!     pub vertex_shader: vertex_shader::Shader
+//! }
+//! 
+//! impl Shaders {
+//!     pub fn load(device: Arc<Device>) -> Result<Self, OomError> {
+//!         Ok(Self {
+//!             vertex_shader: vertex_shader::Shader::load(device)?,
+//!         })
+//!     }
+//! }
+//! ```
+//! 
+//! # Options
+//! 
+//! The options available are in the form of the following attributes:
+//!
+//! ## `#[ty = "..."]`
+//!
+//! This defines what shader type the given GLSL source will be compiled into.
+//! The type can be any of the following:
+//!
+//! * `vertex`
+//! * `fragment`
+//! * `geometry`
+//! * `tess_ctrl`
+//! * `tess_eval`
+//! * `compute`
+//!
+//! For details on what these shader types mean, [see Vulkano's documentation][pipeline].
+//!
+//! ## `#[src = "..."]`
+//!
+//! Provides the raw GLSL source to be compiled in the form of a string. Cannot
+//! be used in conjunction with the `#[path]` attribute.
+//!
+//! ## `#[path = "..."]`
+//!
+//! Provides the path to the GLSL source to be compiled, relative to `Cargo.toml`.
+//! Cannot be used in conjunction with the `#[src]` attribute.
+//! 
+//! [reflect]: https://github.com/vulkano-rs/vulkano/blob/master/vulkano-shaders/src/lib.rs#L67
+//! [cargo-expand]: https://github.com/dtolnay/cargo-expand
+//! [ShaderModule::new]: https://docs.rs/vulkano/*/vulkano/pipeline/shader/struct.ShaderModule.html#method.new
+//! [OomError]: https://docs.rs/vulkano/*/vulkano/enum.OomError.html
+//! [pipeline::shader]: https://docs.rs/vulkano/*/vulkano/pipeline/shader/index.html
+//! [descriptor]: https://docs.rs/vulkano/*/vulkano/descriptor/index.html
+//! [ShaderStages]: https://docs.rs/vulkano/*/vulkano/descriptor/descriptor/struct.ShaderStages.html
+//! [PipelineLayoutDesc]: https://docs.rs/vulkano/*/vulkano/descriptor/pipeline_layout/trait.PipelineLayoutDesc.html
+//! [SpecializationConstants]: https://docs.rs/vulkano/*/vulkano/pipeline/shader/trait.SpecializationConstants.html
+//! [pipeline]: https://docs.rs/vulkano/*/vulkano/pipeline/index.html
 
 #![doc(html_logo_url = "https://raw.githubusercontent.com/vulkano-rs/vulkano/master/logo.png")]
 
 #![recursion_limit = "1024"]
 #[macro_use] extern crate quote;
              extern crate shaderc;
+             extern crate proc_macro;
              extern crate proc_macro2;
-             extern crate syn;
+#[macro_use] extern crate syn;
 
-use std::io::Error as IoError;
 
-use syn::Ident;
-use proc_macro2::{Span, TokenStream};
-use shaderc::{Compiler, CompileOptions};
+use std::env;
+use std::fs::File;
+use std::io::Read;
+use std::path::Path;
 
-pub use shaderc::{CompilationArtifact, ShaderKind};
-pub use parse::ParseError;
+use syn::parse::{Parse, ParseStream, Result};
+use syn::{Ident, LitStr, LitBool};
 
-use parse::Instruction;
-use enums::Capability;
-
+mod codegen;
 mod descriptor_sets;
 mod entry_point;
 mod enums;
@@ -35,295 +182,152 @@ mod spec_consts;
 mod structs;
 mod spirv_search;
 
-pub fn compile(code: &str, ty: ShaderKind) -> Result<CompilationArtifact, String> {
-    let mut compiler = Compiler::new().ok_or("failed to create GLSL compiler")?;
-    let compile_options = CompileOptions::new().ok_or("failed to initialize compile option")?;
+use codegen::ShaderKind;
 
-    let content = compiler
-        .compile_into_spirv(&code, ty, "shader.glsl", "main", Some(&compile_options))
-        .map_err(|e| e.to_string())?;
-
-    Ok(content)
+enum SourceKind {
+    Src(String),
+    Path(String),
 }
 
-pub fn reflect(name: &str, spirv: &[u32], mod_name: &Ident, dump: bool) -> Result<TokenStream, Error> {
-    let struct_name = Ident::new(&name, Span::call_site());
-    let doc = parse::parse_spirv(spirv)?;
+struct MacroInput {
+    mod_ident: Ident,
+    shader_kind: ShaderKind,
+    source_kind: SourceKind,
+    dump: bool,
+}
 
-    // checking whether each required capability is enabled in the Vulkan device
-    let mut cap_checks: Vec<TokenStream> = vec!();
-    for i in doc.instructions.iter() {
-        if let &Instruction::Capability(ref cap) = i {
-            if let Some(cap_string) = capability_name(cap) {
-                let cap = Ident::new(cap_string, Span::call_site());
-                cap_checks.push(quote!{
-                    if !device.enabled_features().#cap {
-                        panic!("capability {:?} not enabled", #cap_string); // TODO: error
-                        //return Err(CapabilityNotEnabled);
+
+/// TODO: Alternative macro syntax:
+/// ```
+///     vulkano_shader! {
+///         mod vs {
+///             Arguments {
+///                 ty: "vertex",
+///                 src: "
+/// #version 450
+///
+/// layout(location = 0) in vec2 position;
+///
+/// void main() {
+/// gl_Position = vec4(position, 0.0, 1.0);
+/// }"
+///             }
+///         }
+///     }
+/// ```
+
+
+impl Parse for MacroInput {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let mut dump = None;
+        let mut mod_ident = None;
+        let mut shader_kind = None;
+        let mut source_kind = None;
+
+        while !input.is_empty() {
+            let name: Ident = input.parse()?;
+            input.parse::<Token![:]>()?;
+
+            match name.to_string().as_ref() {
+                "mod_name" => {
+                    if mod_ident.is_some() {
+                        panic!("Only one `mod` can be defined")
                     }
-                });
+
+                    let mod_name: Ident = input.parse()?;
+                    mod_ident = Some(mod_name);
+                }
+                "ty" => {
+                    if shader_kind.is_some() {
+                        panic!("Only one `ty` can be defined")
+                    }
+
+                    let ty: LitStr = input.parse()?;
+                    let ty = match ty.value().as_ref() {
+                        "vertex" => ShaderKind::Vertex,
+                        "fragment" => ShaderKind::Fragment,
+                        "geometry" => ShaderKind::Geometry,
+                        "tess_ctrl" => ShaderKind::TessControl,
+                        "tess_eval" => ShaderKind::TessEvaluation,
+                        "compute" => ShaderKind::Compute,
+                        _ => panic!("Unexpected shader type, valid values: vertex, fragment, geometry, tess_ctrl, tess_eval, compute")
+                    };
+                    shader_kind = Some(ty);
+                }
+                "src" => {
+                    if source_kind.is_some() {
+                        panic!("Only one `src` or `path` can be defined")
+                    }
+
+                    let src: LitStr = input.parse()?;
+                    source_kind = Some(SourceKind::Src(src.value()));
+                }
+                "path" => {
+                    if source_kind.is_some() {
+                        panic!("Only one `src` or `path` can be defined")
+                    }
+
+                    let path: LitStr = input.parse()?;
+                    source_kind = Some(SourceKind::Path(path.value()));
+                }
+                "dump" => {
+                    if dump.is_some() {
+                        panic!("Only one `dump` can be defined")
+                    }
+                    let dump_lit: LitBool = input.parse()?;
+                    dump = Some(dump_lit.value);
+                }
+                name => panic!(format!("Unknown field name: {}", name))
+            }
+
+            if !input.is_empty() {
+                input.parse::<Token![,]>()?;
             }
         }
+
+        let shader_kind = match shader_kind {
+            Some(shader_kind) => shader_kind,
+            None => panic!("Please provide a shader type e.g. `ty: \"vertex\"`")
+        };
+
+        let source_kind = match source_kind {
+            Some(source_kind) => source_kind,
+            None => panic!("Please provide a source e.g. `path: \"foo.glsl\"` or `src: \"glsl source code here ...\"`")
+        };
+
+        let mod_ident = match mod_ident {
+            Some(mod_ident) => mod_ident,
+            None => panic!("Please provide a mod e.g. `mod: fs` ")
+        };
+
+        let dump = dump.unwrap_or(false);
+
+        Ok(MacroInput { shader_kind, source_kind, mod_ident, dump })
     }
+}
 
-    // writing one method for each entry point of this module
-    let mut entry_points_inside_impl: Vec<TokenStream> = vec!();
-    let mut entry_points_outside_impl: Vec<TokenStream> = vec!();
-    for instruction in doc.instructions.iter() {
-        if let &Instruction::EntryPoint { .. } = instruction {
-            let (outside, entry_point) = entry_point::write_entry_point(&doc, instruction);
-            entry_points_inside_impl.push(entry_point);
-            entry_points_outside_impl.push(outside);
-        }
-    }
+#[proc_macro]
+pub fn vulkano_shader(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+    let input = parse_macro_input!(input as MacroInput);
 
-    let structs = structs::write_structs(&doc);
-    let descriptor_sets = descriptor_sets::write_descriptor_sets(&doc);
-    let specialization_constants = spec_consts::write_specialization_constants(&doc);
-    let ast = quote!{
-        mod #mod_name {
-            #[allow(unused_imports)]
-            use std::sync::Arc;
-            #[allow(unused_imports)]
-            use std::vec::IntoIter as VecIntoIter;
+    let source_code = match input.source_kind {
+        SourceKind::Src(source) => source,
+        SourceKind::Path(path) => {
+            let root = env::var("CARGO_MANIFEST_DIR").unwrap_or(".".into());
+            let full_path = Path::new(&root).join(&path);
 
-            #[allow(unused_imports)]
-            use vulkano::device::Device;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor::DescriptorDesc;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor::DescriptorDescTy;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor::DescriptorBufferDesc;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor::DescriptorImageDesc;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor::DescriptorImageDescDimensions;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor::DescriptorImageDescArray;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor::ShaderStages;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor_set::DescriptorSet;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor_set::UnsafeDescriptorSet;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::descriptor_set::UnsafeDescriptorSetLayout;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::pipeline_layout::PipelineLayout;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::pipeline_layout::PipelineLayoutDesc;
-            #[allow(unused_imports)]
-            use vulkano::descriptor::pipeline_layout::PipelineLayoutDescPcRange;
-            #[allow(unused_imports)]
-            use vulkano::pipeline::shader::SpecializationConstants as SpecConstsTrait;
-            #[allow(unused_imports)]
-            use vulkano::pipeline::shader::SpecializationMapEntry;
-
-            pub struct #struct_name {
-                shader: ::std::sync::Arc<::vulkano::pipeline::shader::ShaderModule>,
+            if full_path.is_file() {
+                let mut buf = String::new();
+                File::open(full_path)
+                    .and_then(|mut file| file.read_to_string(&mut buf))
+                    .expect(&format!("Error reading source from {:?}", path));
+                buf
+            } else {
+                panic!("File {:?} was not found ; note that the path must be relative to your Cargo.toml", path);
             }
-
-            impl #struct_name {
-                /// Loads the shader in Vulkan as a `ShaderModule`.
-                #[inline]
-                #[allow(unsafe_code)]
-                pub fn load(device: ::std::sync::Arc<::vulkano::device::Device>)
-                            -> Result<#struct_name, ::vulkano::OomError>
-                {
-                    #( #cap_checks )*
-                    let words = [ #( #spirv ),* ];
-
-                    unsafe {
-                        Ok(#struct_name {
-                            shader: try!(::vulkano::pipeline::shader::ShaderModule::from_words(device, &words))
-                        })
-                    }
-                }
-
-                /// Returns the module that was created.
-                #[allow(dead_code)]
-                #[inline]
-                pub fn module(&self) -> &::std::sync::Arc<::vulkano::pipeline::shader::ShaderModule> {
-                    &self.shader
-                }
-
-                #( #entry_points_inside_impl )*
-            }
-
-            #( #entry_points_outside_impl )*
-
-            pub mod ty {
-                #structs
-            }
-
-            #descriptor_sets
-            #specialization_constants
         }
     };
 
-    if dump {
-        println!("{}", ast.to_string());
-        panic!("vulkano_shader! rust codegen dumped") // TODO: use span from dump
-    }
-
-    Ok(ast)
-}
-
-#[derive(Debug)]
-pub enum Error {
-    IoError(IoError),
-    ParseError(ParseError),
-}
-
-impl From<IoError> for Error {
-    #[inline]
-    fn from(err: IoError) -> Error {
-        Error::IoError(err)
-    }
-}
-
-impl From<ParseError> for Error {
-    #[inline]
-    fn from(err: ParseError) -> Error {
-        Error::ParseError(err)
-    }
-}
-
-/// Returns the name of the Vulkan something that corresponds to an `OpCapability`.
-///
-/// Returns `None` if irrelevant.
-// TODO: this function is a draft, as the actual names may not be the same
-fn capability_name(cap: &Capability) -> Option<&'static str> {
-    match *cap {
-        Capability::CapabilityMatrix => None,        // always supported
-        Capability::CapabilityShader => None,        // always supported
-        Capability::CapabilityGeometry => Some("geometry_shader"),
-        Capability::CapabilityTessellation => Some("tessellation_shader"),
-        Capability::CapabilityAddresses => panic!(), // not supported
-        Capability::CapabilityLinkage => panic!(),   // not supported
-        Capability::CapabilityKernel => panic!(),    // not supported
-        Capability::CapabilityVector16 => panic!(),  // not supported
-        Capability::CapabilityFloat16Buffer => panic!(), // not supported
-        Capability::CapabilityFloat16 => panic!(),   // not supported
-        Capability::CapabilityFloat64 => Some("shader_f3264"),
-        Capability::CapabilityInt64 => Some("shader_int64"),
-        Capability::CapabilityInt64Atomics => panic!(),  // not supported
-        Capability::CapabilityImageBasic => panic!(),    // not supported
-        Capability::CapabilityImageReadWrite => panic!(),    // not supported
-        Capability::CapabilityImageMipmap => panic!(),   // not supported
-        Capability::CapabilityPipes => panic!(), // not supported
-        Capability::CapabilityGroups => panic!(),    // not supported
-        Capability::CapabilityDeviceEnqueue => panic!(), // not supported
-        Capability::CapabilityLiteralSampler => panic!(),    // not supported
-        Capability::CapabilityAtomicStorage => panic!(), // not supported
-        Capability::CapabilityInt16 => Some("shader_int16"),
-        Capability::CapabilityTessellationPointSize =>
-            Some("shader_tessellation_and_geometry_point_size"),
-        Capability::CapabilityGeometryPointSize =>
-            Some("shader_tessellation_and_geometry_point_size"),
-        Capability::CapabilityImageGatherExtended => Some("shader_image_gather_extended"),
-        Capability::CapabilityStorageImageMultisample =>
-            Some("shader_storage_image_multisample"),
-        Capability::CapabilityUniformBufferArrayDynamicIndexing =>
-            Some("shader_uniform_buffer_array_dynamic_indexing"),
-        Capability::CapabilitySampledImageArrayDynamicIndexing =>
-            Some("shader_sampled_image_array_dynamic_indexing"),
-        Capability::CapabilityStorageBufferArrayDynamicIndexing =>
-            Some("shader_storage_buffer_array_dynamic_indexing"),
-        Capability::CapabilityStorageImageArrayDynamicIndexing =>
-            Some("shader_storage_image_array_dynamic_indexing"),
-        Capability::CapabilityClipDistance => Some("shader_clip_distance"),
-        Capability::CapabilityCullDistance => Some("shader_cull_distance"),
-        Capability::CapabilityImageCubeArray => Some("image_cube_array"),
-        Capability::CapabilitySampleRateShading => Some("sample_rate_shading"),
-        Capability::CapabilityImageRect => panic!(), // not supported
-        Capability::CapabilitySampledRect => panic!(),   // not supported
-        Capability::CapabilityGenericPointer => panic!(),    // not supported
-        Capability::CapabilityInt8 => panic!(),  // not supported
-        Capability::CapabilityInputAttachment => None,       // always supported
-        Capability::CapabilitySparseResidency => Some("shader_resource_residency"),
-        Capability::CapabilityMinLod => Some("shader_resource_min_lod"),
-        Capability::CapabilitySampled1D => None,        // always supported
-        Capability::CapabilityImage1D => None,        // always supported
-        Capability::CapabilitySampledCubeArray => Some("image_cube_array"),
-        Capability::CapabilitySampledBuffer => None,         // always supported
-        Capability::CapabilityImageBuffer => None,        // always supported
-        Capability::CapabilityImageMSArray => Some("shader_storage_image_multisample"),
-        Capability::CapabilityStorageImageExtendedFormats =>
-            Some("shader_storage_image_extended_formats"),
-        Capability::CapabilityImageQuery => None,        // always supported
-        Capability::CapabilityDerivativeControl => None,        // always supported
-        Capability::CapabilityInterpolationFunction => Some("sample_rate_shading"),
-        Capability::CapabilityTransformFeedback => panic!(), // not supported
-        Capability::CapabilityGeometryStreams => panic!(),   // not supported
-        Capability::CapabilityStorageImageReadWithoutFormat =>
-            Some("shader_storage_image_read_without_format"),
-        Capability::CapabilityStorageImageWriteWithoutFormat =>
-            Some("shader_storage_image_write_without_format"),
-        Capability::CapabilityMultiViewport => Some("multi_viewport"),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_bad_alignment() {
-        // vec3/mat3/mat3x* are problematic in arrays since their rust
-        // representations don't have the same array stride as the SPIR-V
-        // ones. E.g. in a vec3[2], the second element starts on the 16th
-        // byte, but in a rust [[f32;3];2], the second element starts on the
-        // 12th byte. Since we can't generate code for these types, we should
-        // create an error instead of generating incorrect code.
-        let comp = compile("
-        #version 450
-        struct MyStruct {
-            vec3 vs[2];
-        };
-        layout(binding=0) uniform UBO {
-            MyStruct s;
-        };
-        void main() {}
-        ", ShaderKind::Vertex).unwrap();
-        let doc = parse::parse_spirv(comp.as_binary()).unwrap();
-        let res = std::panic::catch_unwind(|| structs::write_structs(&doc));
-        assert!(res.is_err());
-    }
-    #[test]
-    fn test_trivial_alignment() {
-        let comp = compile("
-        #version 450
-        struct MyStruct {
-            vec4 vs[2];
-        };
-        layout(binding=0) uniform UBO {
-            MyStruct s;
-        };
-        void main() {}
-        ", ShaderKind::Vertex).unwrap();
-        let doc = parse::parse_spirv(comp.as_binary()).unwrap();
-        structs::write_structs(&doc);
-    }
-    #[test]
-    fn test_wrap_alignment() {
-        // This is a workaround suggested in the case of test_bad_alignment,
-        // so we should make sure it works.
-        let comp = compile("
-        #version 450
-        struct Vec3Wrap {
-            vec3 v;
-        };
-        struct MyStruct {
-            Vec3Wrap vs[2];
-        };
-        layout(binding=0) uniform UBO {
-            MyStruct s;
-        };
-        void main() {}
-        ", ShaderKind::Vertex).unwrap();
-        let doc = parse::parse_spirv(comp.as_binary()).unwrap();
-        structs::write_structs(&doc);
-    }
+    let content = codegen::compile(&source_code, input.shader_kind).unwrap();
+    codegen::reflect("Shader", content.as_binary(), &input.mod_ident, input.dump).unwrap().into()
 }
